requirements.txt
torchvision>=0.7.0
pillow>=7.2.0
numpy>=1.18.5
scipy>=1.5.0
scikit-learn>=0.23.1
matplotlib>=3.2.2
h5py>=2.10.0
setuptools>=49.2.0
readline>=8.0
pyzmq>=19.0.1